requirements.txt
# AWS SDK
boto3==1.34.0
botocore==1.34.0

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# File Handling
python-multipart==0.0.6
Pillow==10.1.0

# Validation
pydantic==2.5.0

# CORS
fastapi-cors==0.0.6

# Utilities
uuid==1.30

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Optional: Database (if needed later)
# sqlalchemy==2.0.23
# alembic==1.13.0